refactor: Organiza tipagens e consultas comuns
