+++ src/schedules/utils/select-schedule.prisma-sql.ts
@@ -1,0 +1,49 @@
+const userSelect = {
+  user: {
+    select: {
+      id: true,
+      name: true,
+      email: true,
+    },
+  },
+};
+
+const studentSelect = {
+  student: {
+    select: {
+      enrollment: true,
+      ...userSelect,
+    },
+  },
+};
+
+const scheduleSelectPrismaSQL = {
+  id: true,
+  id_status: true,
+  start: true,
+  end: true,
+  ...studentSelect,
+  monitor: {
+    select: {
+      id: true,
+      ...studentSelect,
+      subject: {
+        select: {
+          id: true,
+          name: true,
+          course: {
+            select: {
+              id: true,
+              name: true,
+            },
+          },
+        },
+      },
+      responsible_professor: {
+        select: { ...userSelect },
+      },
+    },
+  },
+};
+
+export default scheduleSelectPrismaSQL;
